requirements.txt
faiss-cpu
faiss-gpu
torch
torchvision
streamlit
tqdm
python-slugify
opendatasets
azure-storage-blob
streamlit-cropper